Scripts: More error-handling for swaylock

+++ scripts/bin/scripts/swaylock_multimonitor-laptop.sh
@@ -2,15 +2,27 @@
 #set -x
 
 LOCKARGS=""
-for OUTPUT in $(swaymsg -t get_outputs | jq -r '.[] | select(.active == true) | .name')
+NORM=""
+MONITORS=$(swaymsg -t get_outputs | jq -r '.[] | select(.active == true) | .name')
+for OUTPUT in $MONITORS
 do
     IMAGE=/tmp/$OUTPUT-lock.png
     grim -o "$OUTPUT" "$IMAGE"
     convert "$IMAGE" -scale 10% -scale 1000% "$IMAGE"
-    LOCKARGS="${LOCKARGS} --image ${OUTPUT}:${IMAGE}"
-    IMAGES="${IMAGES} ${IMAGE}"
+    if [ -f $IMAGE ]
+    then 
+        LOCKARGS="${LOCKARGS} --image ${OUTPUT}:${IMAGE}"
+        IMAGES="${IMAGES} ${IMAGE}"
+    else
+        LOCKARGS="-c 000000"
+        NORM=1
+    fi
 done
 
 #dbus-send --print-reply --dest=org.mpris.MediaPlayer2.spotify /org/mpris/MediaPlayer2 org.mpris.MediaPlayer2.Player.Stop
-swaylock $LOCKARGS
-rm $IMAGES
+swaylock -f $LOCKARGS
+
+if [ -z  "$NORM" ]
+then
+    rm $IMAGES
+fi
